[barray] adopt the barray module
This is WIP, but I think we can do a quick review for it.

A few things in this change:
1. I added the iterator module to replace the sequence module. The OCaml
community seems have not reached an aggrement on how functional
iterator would look like, we've got `gen`, `seq`, `Core.Sequence` etc. I
picked this because it is purely functional and have a monadic
interface, while `Sequence` is side-effect based. Dave suggested we use
this as well.
2. Adopted the barray module, added a Collection interface

Things to do:
1. add unit tests for barray